🧹 Refactor Class Locations
- Moves classes to packages that made more sense.
