+++ src/main/java/com/marcusslover/plus/lib/common/interfaces/ThrowableSupplier.java
@@ -1,4 +1,4 @@
-package com.marcusslover.plus.lib.exception;
+package com.marcusslover.plus.lib.common.interfaces;
 
 @FunctionalInterface
 public interface ThrowableSupplier<T extends Throwable, V> {
